Cleanup the tests
